Altered two directories in response to valgrind timing out on thread-ring.  The timeout is a result of limiting the number of threads for valgrind and since the number of threads is very important to the thread-ring program, I decided to skip the thread-ring directory when running valgrind, rather than change the run details.  I updated the README file appropriately.
However, I chose a different path for the comparison of our different task
creation models in test/parallel/taskCompare.  While the point of this directory
is to show the performance of different task creation models, it seems sensible
to at least assure ourselves that thread-ring does not leak memory.  Thus, I
changed the expected output and .good files used to use less threads than the
cap imposed on valgrind runs.  This test directory runs in a reasonable amount
of time (roughly twenty minutes for twelve tests) under valgrind.  I did not
feel the need to reflect this change over to test/studies/shootout/thread-ring,
as the settings there allow it to run in a few minutes under normal testing
settings and it is better to check the accuracy for a higher number of threads
since the shootout size is so large (503).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21804 3a8e244f-b0f2-452b-bcba-4c88e055c3ca