Improving handling of CHPL_COMPILER on Crays when no PrgEnv loaded
Prior to this commit, when running the compiler script on cray-x*
systems without a PrgEnv loaded, the compiler string would be set
to 'cray-prgenv' rather than 'cray-prgenv-foo' which is what we
expect when PrgEnv-foo was loaded.  This has confused me twice
this week because it doesn't seem inherently problematic until
you really look at it.  From what I can tell, this has been the
behavior since the beginning of time.

To try and improve this situation, this commit implements two
small changes:

1) emits a warning in this event.

2) makes the string 'cray-prgenv-none' just to make it stand out
   a bit more clearly if one were to miss the warnings.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21808 3a8e244f-b0f2-452b-bcba-4c88e055c3ca