Filing some future for != and == of empty records
Empty record == and != always return true or false respectively,
regardless of the type of the second argument (int in these cases).

This happens because we build the functions with the second argument
as dtAny to handle sub-records, and then they later get instantiated
during resolution with the actual type.  The body of the function is
empty other than the return statement because the records have not
fields, so it compiles and runs.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21812 3a8e244f-b0f2-452b-bcba-4c88e055c3ca