Bug fix and minor tweaks to modules/standard/AdvancedIters.chpl.
[committing on behalf of Ben, plus minor tweaks]

* The dynamic iterator used to have all the work done in the (single)
current task when the number of chunks is less than the number of tasks.

This commit makes the current task handle all work when it all fits in
one chunk. Otherwise, the number of tasks created is min(tasks to
satisfy work, number of user specified tasks).

Note that having the current task handle all work has (and has had)
the benefit of avoiding the creation and use of a sync variable,
splitLock$, and other overhead.

** TODO: apply the same to guided and adaptive iterators?

* Do not report "serial execution because there is not enough work"
when there is only a single task. This is for consistency with
other leader iterators that stick to a single task when there is little
work but do not print the warning.

** The above report is enabled, though, when debugAdvancedIters is enabled.

* Converted a halt() to a compilerError().

* Asserted that the 'chunkSize' argument is positive and
halted unless 'numTasks' is positive or 0.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21813 3a8e244f-b0f2-452b-bcba-4c88e055c3ca