Added a test that verifies the correctness of an idiom in ChapelLocale.chpl.
This test verifies that updating a pragma-"private" global variable
within a coforall+on works as expected, i.e., updates the copy
of that variable that is private to that locale.
This idiom occurs in modules/internal/ChapelLocale.chpl for 'rootLocale'.

I am adding this test now to ensure that correctness is preserved
by the upcoming "capture for tasking" language change.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21820 3a8e244f-b0f2-452b-bcba-4c88e055c3ca