Adding 'cleanall' target to Makefiles
[reviewed by vass and greg]

This commit adds a new 'cleanall' target that sits between 'clean' and
'clobber' (in terms of the amount that it cleans up) to our Makefiles.
The current definition of 'clean' only removes things directly related
to the current configuration -- for example, it will only clean the
subdirectory under any 'gen' directory in compiler/ or runtime/
corresponding to the current printchplenv configuration.  This new
'cleanall' target will remove the 'gen' directory itself, removing
all configurations.  

Note that, like clobber, it only removes such directories for the
paths that the Makefiles traverse into in their normal recursive
crawl.  Thus, if you're configured for gasnet, it won't descend into
runtime/src/comm/none/...  If we wanted to make these rules even more
independent of the current settings (which is probably a good goal),
we'd need to switch away from our current Makefile-based traversal and
use something more agressive like a find command.  I didn't feel
confident enough about that strategy to go down this path yet, but we
may want to do it in the future.  (In particular, my one reluctance is
that I like that the current approach permits each directory to
specify what should be cleaned up under it recursively; it may be that
we could also broaden the subdirectories that are considered for
cleanall/clobbering to include all subdirectories within the Makefile
structure without switching to a find, but I didn't investigate that
here).

For the compiler/ and runtime/ directories, 'cleanall' differs from
'clobber' only in that 'clobber' will remove the top-level bin/ and
lib/ directories and 'cleanall' (like 'clean') leaves the final
artifacts of a build intact.  The bigger difference is in the
third-party directories where 'cleanall' will remove the entire build/
directory whereas 'clobber' removes both the build and the install
directories.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21824 3a8e244f-b0f2-452b-bcba-4c88e055c3ca