Triage updates for the last week through today.
REGRESSIONS

* A bunch of compilation timeouts started 9/6.
Assigned to Elliot as a first approximation.

These were a bit hard to attribute to specific configurations -
each config has its own set of timeouts slightly different from
other configs. So I just listed most of them at the top of the file
under "general" and "linux64".

As an exception, I put a full list of timeouts for the
"no-local" and "valgrind" configurations.

* Valgrind timeouts for thread-ring variants have been resolved.

* Added a section on build chaos on chpbld01.
Brad has been working on those.

* More segfaults for PGI and Intel.

* Removed test_dot_locale_error from a "Comm diags mismatch" list,
as it does not count comms.

TRIAGE

* New dates for a couple of known intermittent timeouts.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21834 3a8e244f-b0f2-452b-bcba-4c88e055c3ca