Removing an unused variable
[trivial]

This is causing build failures in developer mode for gcc/4.8.0.
Note that I cleaned this up in the simplest, but maybe not
most elegant, way possible.  I wanted to get it cleaned up
before tonight's tests ran, but am open to someone (it's Tom's
code) cleaning it up more after the fact.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21836 3a8e244f-b0f2-452b-bcba-4c88e055c3ca