Modify testEnv to set all enviro vars from printchplenv
[reviewed by Sung]

Some tests, such as test/io/extern/ferguson/externblock, have a
SKIPIF file that use a Chapel configuration environment variable
that is often not set. In that case, the CHPL_LLVM environment
variable is automatically set by chplenv/llvm if it detects
a built LLVM, but start_test/testEnv doesn't know about that. This
patch modifies testEnv to set all values from printchplenv so
that they can be used in .skipif or SKIPIF files.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21841 3a8e244f-b0f2-452b-bcba-4c88e055c3ca