Fix memory trashing bug when CHPL_MAKE or --make is set
[reviewed by Brad]

The command-line parsing program had type S256 for CHPL_MAKE,
even though it was declared as a const char* pointer. That meant that
it would just overwrite the next several bytes in memory
if the variable was ever set. We added a 256-byte long
static variable, makeArgument, and asked the command-line parsing
to call handleMake, which just sets CHPL_MAKE to point to
makeArgument. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21842 3a8e244f-b0f2-452b-bcba-4c88e055c3ca