Be smarter about rebuilding the compiler based on configuration
[reviewed by Brad]

Earlier work properly stored the configuration in the gen
directories storing the .o files in compiler/. However,
since e.g. bin/linux64/chpl already exists, flipping between
configurations (for example enabling and then disabling LLVM)
does not cause a compiler rebuild. This patch adds a
file, .built-for-CONFIG-HERE and some simple make rules.
First, the Chapel binary has a dependency on the .built-for
file for the current configuration. Second, the rule to
build the .built-for file removes any other .built-for files
and then outputs one storing the current configuration. With
this change I have been able to flip between
CHPL_LLVM=none and CHPL_LLVM=llvm and get the compiler
to re-link each time.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21847 3a8e244f-b0f2-452b-bcba-4c88e055c3ca