Add localeid method to the locale class
[reviewed by Greg]

This commit adds a method called localeid to the locale class that
returns the locale id.  Locale models must override this function.
The locale id type is defined by the Locale Model.

Background:

We have been using id as a unique integer identifier to mean what we
now refer to as the node id.  Since the hierarchical locales merge did
not utilize sublocale information, simply having the node id available
worked just fine.  As we have been adding support for sublocales in a
NUMA locale model, we need the actual locale id, which for NUMA will
include node id and sublocale id.

Based on discussions, it seem that we may want to eventually change id
to something else (e.g., nodeid, nid, uid) and use id for the locale
id.  For this mod, I simply added localeid, with the argument that the
hierarchical locales support will not be heavily used for a while and
so maintaining id as is would be fine, plus any code using id won't
have to change yet (e.g., bulk comm opt).  We can make a better
informed decision when we have more experience with hierarchical
locales and not have to possibly change things twice.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21851 3a8e244f-b0f2-452b-bcba-4c88e055c3ca