Make fVerify false by default.  Fix typo in nightly script. [trivial]
In post-pass checking, verifications inside "if (fVerify) ..." clauses are now
skipped by default.  Toss the --verify flag on the compile line or set
CHPL_VERFIY=1 in your environment to enable these checks.

Nightly verification test runs did not have a distinctive subject line in the
cron emails.  This was due to a cut-and-paste error on line 318 of
util/cron/nightly.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21854 3a8e244f-b0f2-452b-bcba-4c88e055c3ca