Update .good output for memory/shannon/memstatPrint.
[Reviewed by sungeun, conversationally.]

There were a couple of things causing memstatPrint to fail.  First,
since 5/28/09 it hasn't worked with comm=gasnet and CHPL_GASNET_SEGMENT
set to "fast".  This commit adds a .skipif to make us skip the test in
that situation.  Second, my r21838 commit changed the output, because it
changed some untracked allocations into tracked ones.  As a result the
.no-local.good file needed updating and we also needed to add a .good
file specifically for comm-gasnet.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21857 3a8e244f-b0f2-452b-bcba-4c88e055c3ca