When CHPL_LOCALE_MODEL is "numa" generate wide references and outlined on fns
[Reviewed by Sung]

The "numa" locale model needs wide references and outlined on functions even
when compiling --local.  The wide references are used to track the sublocales
of references.  It needs outlined on functions so that "on subloc do" can go
through the LocaleModel interface to get moved onto a sublocale.

Since the "numa" module is not yet available, I tested this by generating wide
references and outlined on functions for the "flat" model instead.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21875 3a8e244f-b0f2-452b-bcba-4c88e055c3ca