Adjusted what's displayed in the error message "non-lvalue actual is passed to ... formal".
This is a follow-up to r21735, requested by Brad.

* Do not mention the name of the formal in the error message
if the callee is in an internal module. In Brad's words, revealing
"an "internal" argument name which seems unfortunate in that (1) it
won't mean much to the user and (2) it makes the test sensitive to
something that I don't think we want to commit to."

  I suppress the name of the formal if the callee is in a standard
module, by the same argument and also because it is done this way
in insertLineNumber(). BTW maybe we should hoist that check into
a bool revealToUser(FnSymbol*) (which would depend on --devel)
or into revealToNonDeveloper(FnSymbol*) (which wouldn't).

* OTOH show the name of the function that is being invoked. This will
probably be more helpful to the programmer than the name of the formal,
especially when the source line with the error has several function calls.

* Use single quotes instead of double quotes. In Brad's words, "The
use of double-quotes around the formal argument name looks awkward to
me, and to my memory/thinking, I don't think we tend to use single
quotes in other error message contexts like this. [...] if there's a
precedent for [double quotes] (in our compiler or other notable ones),
I'm open to being convinced otherwise."

  I found a mixed use of single and double quotes in the compiler and
modules, so for simplicity I went with that request.

* When the formal's name is not included in the error message,
I switch it from "formal 'myFormalName'" to "a formal".

* Include "()" after the callee's name, unless it looks like an operator.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21883 3a8e244f-b0f2-452b-bcba-4c88e055c3ca