Ditto r21900
This should fix numa+qthreads.

For those wondering if we're crazy, there's a perfectly logical
explanation.  The problem was that Greg had a couple of 'SubLoc'
typos, but since he tested his change before NUMA and assumed I had
tested the qthreads part, he didn't notice those typos.  In the
meantime I was testing qthreads with 'SubLoc' since that's was what it
was in task-qthreads.c.  I think we got it straightened out now.

Greg has volunteered for the clean up task of standardizing on
subloc/Subloc.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21901 3a8e244f-b0f2-452b-bcba-4c88e055c3ca