Removing locale model from the compiler's object file paths
[reviewed by sungeun, gbt]

The hierarchical locale team noted that things were getting
recompiled too often when they switched between modes.  It
turns out that this was because the hierarchical model was
being used to create the gen/ directory name for the compiler
objects.  The compiler binary does not rely on the locale
model, though, so this commit removes that piece from the
uniquifying string.

There was a question as to whether the pointer representation
required a different compiler binary as well, but we weren't
confident enough in the answer to change that.  (Michael?)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21903 3a8e244f-b0f2-452b-bcba-4c88e055c3ca