Class optimization for LICM
[reviewed by sung] 

In a previous commit, I made it so that defs of a class's (I'm going to refer to
them as aggregate) field were seen as a def to the class and to the field.
The reason for that is because in the case of a record or a union, if a field
is changed, then the aggregate has also changed and you need to note that. In
the case of an aggregate that is actually a class though, the aggregate is
represented by a pointer. The pointer itself doesn't change with a def to a
field. 

This update removes a todo that I should have removed last commit, doesn't mark
defs to class's in the case that a class's field has changed, and cleans up
that code a bit. Note that this is more of an optimization than a bug fix. The
previous code was still correct, just more conservative than it needed to be. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21907 3a8e244f-b0f2-452b-bcba-4c88e055c3ca